in seed tests (32 bit especially) deal with dupe seeds
We expect the RNG to produce a certain number of collissions just
by producing duplicate seeds. This effect is especially pronounced
with 32 bit seeds.

So we dedupe the tests by ignoring the duplicate seeds, so that any
duplicates that TestHashList sees are from the hash function itself,
not duplicate seeds.

We possibly could achieve the same result with a different type RNG
strategy for the seeds. But this seemed like a better option.