Add a CtrStream "test" and clean up some includes at the same time
This runs a hash in "counter mode", hashing a key of a given size
(currently 8 bytes) which is incremented each time, and then printing
the output to STDOUT. This allows the output to be fed to something
like Dieharder to test the stream as an RNG.

Run like this a "strong" hash function should pass most if not any RNG
test it is thrown up against.

If the counter rolls over then we change the seed according to a new
value (chosen "randomly" from a seeded RNG.)

A future patch will add support for controlling the initial hash seed
via the seed passed to the RNG, and the length of key buffer, and maybe
its initial fill.