add the Lua 53 hash - and a variant that processes the ALL the key
The Lua 53 hash is really low quality, it does not process the full
key, and uses a very poor mixing function as well. Since we know that
skipping bytes in the key is going to make our tests fail I have also
added a variant that uses the same mixing function but processes the
whole key. It is not a very good hash function at all.