get rid of pfHash (finally)
Everything now caters to the new api with seeding and hashing
split up, and everything uses hashfunc<> objects, and not pfHash
pointers to do the hashing. This means that when testing new hash
functions they dont have to worry about supporting the older uint32_t
seed api.

One of the casualties of this is the Verification code stuff, which
was designed to detect if the test framework or the hashes were
behaving differently on different architectures. IMO if this is still
needed it can be added back via the hashfunc<> object interface. So
for now I have chainsawed the Vcode stuff.