Add a test that breaks crc32 based hash functions
Some functions try to gain speed by using crc32 intrinsics. The problem
with this is that it becomes trivial to generate arbitrary number of
collisions for such hashes. This test will break any naive use of crc32,
but anything with a remotely good round function will have no problem
with this test.