increase number of expected collisions for CRC tests
The stats we use to decide if we seen too many collisions
works much better if we expect 100 collisions than if we expect 1.
If you expect 1.19 collision and you get 3 is it an issue? On the
other hand, if you expected 100 and you get 300 then surely the situation
is more clear. I think we could use a more solid technique to decide
these things, and calculate a P-value, but of course any strategy we
choose is going to benefit from more samples.

Also there is the interesting issue of rounding here. The math says
we could have a fractional number of collisions, but collisions are
discrete quantities. You can't have 2.34 collisions. You can have 2,
or 3, but not in between. Anyway. I digress...