extend AppendedZeros test
Originally we compared len k to len k+1, and did not test the zero
seed. Now we test len k to len k+i for i in 1,2,4,8,16, for both
the zero seed and a non-zero seed. This allows us to detect block
hashes which do not naturally escape the zero state, but do mix in
a partial length at the end.