added HighwayHash (as library)
See #28
Just don't know how could just use the avx2 or sse41
variant or the full portable solution. This has the CPUID overhead.

Similarily to SipHash this claims to be fast (5x faster than SipHash)
and is in fact 5x faster, which still makes it the second slowest
tested quality hash.