add new Sanmayce FNV1A_Totenschiff_v1
See GH #73, 32-bit hash, but 64-bit architecture only.

This is a very bad hash.
It fails the AppendedZeroesTest, which makes it trivial to
exploit (just add \0's somewhere in the key, and it will collide),
it's slower than FNV1a-YoshimitsuTRIAD, and it's a poor version of my
FNV2. first wordwise, than per byte, just the last step is better.