FEATURE: Add a significantly improved collision estimation function
This uses a justified bound for using p(collisions >= 1) as an
estimator for E[number of collisions], which lets it use the
known-good approximation for hash counts under that limit.

Also improve --EstimateNbCollision test output, expand its list of
hash counts and bit widths to cover more values that are used in the
tests, and add exact true values along with a utility to compute them
via arbitrary-precision libraries.