PERFORMANCE: Test only 32-bit seeds for hashes that don't use Hash_Seed_init()
Because of how the hash() wrapper function is declared, seeds going
through it can only be 32-bits, no matter what happens inside of it
(e.g. casting back out to 64 bits before passing it to the hash). So,
*only* seeds going through Hash_Seed_init() can be wider than 32
bits. This patch disables invalid/redundant seed testing. This could
be reverted when and if the hash() wrapper changes.