CRITICAL FIX: Single collisions with tiny expected counts should fail
This fixes a serious bug in bad seed finding for 32-bit hashes, and
possibly other places. Previous to this, a single actual collision
with a very small expected collision count (<0.1) would NOT trigger a
test failure. The main BadSeed test only passes a handful of seeds to
TestHashList, so failures could be missed!!

To see this, try this patch without this fix, and run "SMHasher
--test=BadSeeds Spooky32":