umash: update to new implementation
The algorithm changed in May 2021 (and has since been frozen) to
improve the latency on short inputs and the potential throughput
of the 128-bit fingerprint.

Adding `umash_long.inc` next to `umash.c` compiles in
throughput-optimised routines for long inputs. We can also
remove `sse2neon.h`: the implementation comes with native
support for neon with crypto extension.

All in all, there should be no change in strength, and no further
change to the function itself.  Throughput is within 20% of XXH3
and HalftimeHash256 on a 2 GHz EPYC 7713.

TESTED=on x86-64, and CI builds.